Renamed app > router.js. VirtualLottery in router.
